+++ demo/src/config/apps.config.ts
@@ -11,10 +11,12 @@ export const appsConfig = new AppsConfig({
 	apps: [
 		fileExplorer.setName("Files")
 			.setDescription("Browse and manage your virtual files on ProzillaOS.")
-			.setIconUrl("/assets/apps/icons/file-explorer.svg"),
+			.setIconUrl("/assets/apps/icons/file-explorer.svg")
+			.setShowDesktopIcon(true),
 		terminal.setName("Commands")
 			.setDescription("A command line tool inspired by the Unix shell that runs entirely in your browser using ProzillaOS. Allows you to interact and manipulate the virtual drive and run silly commands.")
-			.setIconUrl("/assets/apps/icons/terminal.svg"),
+			.setIconUrl("/assets/apps/icons/terminal.svg")
+			.setShowDesktopIcon(true),
 		textEditor.setName("Notes")
 			.setDescription("Text editor for reading and writing text documents in a virtual file system using ProzillaOS.")
 			.setIconUrl("/assets/apps/icons/text-editor.svg"),
@@ -25,7 +27,8 @@ export const appsConfig = new AppsConfig({
 			.setIconUrl("/assets/apps/icons/media-viewer.svg"),
 		browser.setName("Browser")
 			.setDescription("Browse the internet.")
-			.setIconUrl("/assets/apps/icons/browser.svg"),
+			.setIconUrl("/assets/apps/icons/browser.svg")
+			.setShowDesktopIcon(true),
 		calculator.setName("Maths")
 			.setDescription("Simple calculator app.")
 			.setIconUrl("/assets/apps/icons/calculator.svg")