+++ packages/apps/app-center/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @prozilla-os/app-center
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @prozilla-os/core@1.3.7
+
 ## 1.0.4
 
 ### Patch Changes