+++ packages/apps/settings/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @prozilla-os/settings
 
+## 1.1.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @prozilla-os/core@1.3.7
+  - @prozilla-os/file-explorer@1.1.12
+
 ## 1.1.11
 
 ### Patch Changes